+++ src/client/java/modToolkit/client/ModToolkitDataGenerator.java
@@ -1,9 +1,9 @@
-package emcMod.client;
+package modToolkit.client;
 
 import net.fabricmc.fabric.api.datagen.v1.DataGeneratorEntrypoint;
 import net.fabricmc.fabric.api.datagen.v1.FabricDataGenerator;
 
-public class EmcModDataGenerator implements DataGeneratorEntrypoint {
+public class ModToolkitDataGenerator implements DataGeneratorEntrypoint {
 
     @Override
     public void onInitializeDataGenerator(FabricDataGenerator fabricDataGenerator) {