+++ src/main/java/modToolkit/ModToolkit.java
@@ -1,8 +1,8 @@
-package emcMod;
+package modToolkit;
 
 import net.fabricmc.api.ModInitializer;
 
-public class EmcMod implements ModInitializer {
+public class ModToolkit implements ModInitializer {
 
     @Override
     public void onInitialize() {