+++ changelog.md
@@ -1,6 +1,16 @@
 # Changelog
 Current version: `v1.0.0`
 
+## v1.1.0 - 07/05/2025
+- Grouped all commands (no more `/seefollow` and `/removefollow`, but `/follow list` and `/follow remove`)
+- Changed `/tools` to `/follow show <on/off>`
+- Unblocked all commands from the `/tools` infrastructure. `/follow hide` will just hide/show the colors of the followed players.
+- Follow and todo data is properly saved now, so will not reset when you close minecraft.
+- Fixed some minor bugs in the todo list.
+- Updated command.md to view all commands.
+- Changed the old name `EmcMod` to `Mod Toolkit` in the entire codebase.
+- Added icon to the mod.
+
 ## v1.0.0 - 06/05/2025
 Initial release of the project. Includes the following features:
 - `Follow` system to follow other players